requirements_minimal.txt
streamlit
scikit-learn
opencv-python-headless
pandas
numpy
joblib
matplotlib
seaborn
Pillow
